requirements.txt
django == 3.1.5
pillow
django-admin-interface
django-rest-framework